backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
anthropic>=0.40.0
bcrypt>=4.1.3
python-multipart>=0.0.9
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
PyPDF2>=3.0.0
pycryptodome>=3.20.0
easyocr>=1.7.0
pymupdf>=1.24.0
langchain-text-splitters>=0.2.0
numpy>=1.26.0
google-generativeai>=0.8.0
groq>=0.11.0
authlib>=1.3.0
gliclass>=0.1.3
